+++ components/dashboard/statuses.vue
@@ -16,7 +16,7 @@
           <span>Détail</span>
           <ul>
             <li
-              v-for="{total, id} of statuses.slice(0, 3)"
+              v-for="{total, id} of translatedStatuses.slice(0, 3)"
               :key="id"
             >
               {{ id }} (<strong>{{ total }}</strong> course(s))
@@ -26,7 +26,7 @@
       </ul>
     </template>
     <template
-      v-if="statuses.length"
+      v-if="translatedStatuses.length"
       #PIE
     >
       <pie-chart :chart-data="pieStatuses" />
@@ -80,13 +80,75 @@ export default {
     ridesFailed: makeRidesComputedFrom(FAILED_STATUSES),
     ridesCanceled: makeRidesComputedFrom(CANCELED_STATUSES),
     ridesInProgress: makeRidesComputedFrom(IN_PROGRESS_STATUSES),
+    translatedStatuses() {
+      return this.statuses.map((status) => {
+        let id;
+        switch (status.id) {
+          case DRAFTED:
+            id = 'En création';
+            break;
+          case CREATED:
+            id = 'Créée';
+            break;
+          case VALIDATED:
+            id = 'Validée';
+            break;
+          case REJECTED_BOUNDARY:
+            id = 'Refusée - Hors-limite';
+            break;
+          case REJECTED_CAPACITY:
+            id = 'Refusée - Capacité';
+            break;
+          case ACCEPTED:
+            id = 'Acceptée par le chauffeur';
+            break;
+          case DECLINED:
+            id = 'Rejetée par le chauffeur';
+            break;
+          case STARTED:
+            id = 'Commencée';
+            break;
+          case WAITING:
+            id = 'En attente';
+            break;
+          case IN_PROGRESS:
+            id = 'En cours';
+            break;
+          case DELIVERED:
+            id = 'Terminée';
+            break;
+          case CANCELED:
+            id = 'Annulée';
+            break;
+          case CANCELED_TECHNICAL:
+            id = 'Annulée - raisons techniques';
+            break;
+          case CANCELED_REQUESTED_CUSTOMER:
+            id = 'Annulée - initiative du passager';
+            break;
+          case CANCELED_CUSTOMER_OVERLOAD:
+            id = 'Annulée - passager trop chargé';
+            break;
+          case CANCELED_CUSTOMER_MISSING:
+            id = 'Annulée - passager manquant';
+            break;
+          default:
+            id = 'Statut manquant';
+            break;
+        }
+        return {
+          ...status,
+          id,
+        };
+      });
+    },
     ridesTotal() {
       return this.statuses.reduce((acc, { total }) => (acc + total), 0);
     },
     pieStatuses() {
       return {
-        labels: this.statuses.map(({ id }) => id),
-        datasets: [{ ...this.backgroundColors, data: this.statuses.map(({ total }) => total) }],
+        labels: this.translatedStatuses.map(({ id }) => id),
+        datasets: [{ ...this.backgroundColors, data: this.translatedStatuses.map(({ total }) => total) }],
       };
     },
   },