+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Fix infosec issue with X-Frame header
 * Fix arrow on calendar tooltip
 * Fix links to update items in lists
+* Fix status translation in dashboard
 * Fix ratings labels
 * Fix minor bug on list.vue
 ## Version 1.4.3